server/requirements.txt
Django
djangorestframework
python-dotenv
python-telegram-bot
